+++ Changelog.md
@@ -29,6 +29,7 @@
 * Add a white background to images shown in the lightbox [#4475](https://github.com/diaspora/diaspora/issues/4475)
 * Refactor getting_started page, test if facebook is available, fix [#4520](https://github.com/diaspora/diaspora/issues/4520)
 * Avoid publishing empty posts [#4542](https://github.com/diaspora/diaspora/pull/4542)
+* Strip markdown from page title when loading the singe post view
 
 ## Features
 * Add oEmbed content to the mobile view [#4343](https://github.com/diaspora/diaspora/pull/4353)